api/models/ratings.py
from sqlalchemy import Column, Integer, Float, String, ForeignKey, DateTime
from sqlalchemy.orm import relationship
from database import Base
import datetime

class Rating(Base):
    __tablename__ = "ratings"

    id = Column(Integer, primary_key=True, index=True)
    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
    content_id = Column(Integer, ForeignKey("content.id"), nullable=False)
    rating = Column(Float, nullable=False)
    review = Column(String, nullable=True)
    created_at = Column(DateTime, default=datetime.datetime.utcnow)

    user = relationship("User", back_populates="ratings")
    content = relationship("Content", back_populates="ratings")

    def __repr__(self):
        return f"<Rating id={self.id} user_id={self.user_id} content_id={self.content_id} rating={self.rating}>"
