requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Cors==3.1.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.28.0
psycopg2-binary==2.9.3
